SUMMARY:   clean up a few Sun compile warnings
     BUGS:      none

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

     built on Sun, cleaned up some mostly minor compile issues.

(checklist version 2007-06-20)

+++ packages/fei/base/fei_BCManager.cpp
@@ -368,9 +368,8 @@ int BCManager::consolidateBCs()
 
     const BCRecord** listdata = &consolidatedList[0];
     int insertPoint = -1;
-    int index = snl_fei::binarySearchPtr<BCRecord>(lastBC,
-						   listdata,
-				      consolidatedList.size(), insertPoint);
+    int index = snl_fei::binarySearchPtr<BCRecord>(lastBC, listdata,
+				      (int)consolidatedList.size(), insertPoint);
 
     if (index >= 0) {
       //lastBC specifies the same nodeID/fieldID pair as a BCRecord that's
